nagme/requirements.txt
argh==0.26.2
arrow==0.13.1
Babel==2.6.0
certifi==2019.3.9
chardet==3.0.4
coverage==4.5.3
dj-database-url==0.5.0
Django==1.11.17
django-bootstrap3==11.0.0
django-dotenv==1.4.2
django-dramatiq==0.6.0
django-forms-bootstrap==3.1.0
django-phonenumber-field==2.2.0
django-timezone-field==3.0
dramatiq==1.5.0
gevent==1.4.0
greenlet==0.4.15
gunicorn==19.9.0
idna==2.8
mock==2.0.0
model-mommy==1.6.0
pathtools==0.1.2
pbr==5.1.3
phonenumbers==8.10.6
pika==0.13.1
Pillow==5.4.1
prometheus-client==0.2.0
psycopg2-binary==2.7.7
PyJWT==1.7.1
PySocks==1.6.8
python-dateutil==2.8.0
pytz==2018.9
PyYAML==5.1
redis==3.2.1
requests==2.21.0
six==1.12.0
twilio==6.25.2
urllib3==1.24.1
watchdog==0.8.3
watchdog-gevent==0.1.0
whitenoise==4.1.2
